# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SlideBanai API",
    "description": "AI-Powered Presentation Platform with Analytics & Collaboration",
    "version": "1.0.0",
    "contact": {
      "name": "SlideBanai Support",
      "url": "https://slidebanai.com/support",
      "email": "support@slidebanai.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.slidebanai.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
      "Presentation": {
        "type": "object",
        "required": ["id", "user_id", "title", "status"],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Q1 2026 Product Launch"
          },
          "description": {
            "type": "string",
            "example": "Comprehensive product launch presentation"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "in_progress", "published"],
            "example": "draft"
          },
          "slides_count": {
            "type": "integer",
            "example": 12
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatePresentationRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "My Presentation"
          },
          "description": {
            "type": "string",
            "example": "A great presentation"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "in_progress", "published"],
            "example": "draft"
          },
          "slides_count": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "Slide": {
        "type": "object",
        "required": ["id", "presentation_id", "slide_number"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "presentation_id": {
            "type": "integer"
          },
          "slide_number": {
            "type": "integer",
            "example": 1
          },
          "content": {
            "type": "string",
            "example": "# Welcome to SlideBanai"
          },
          "background_color": {
            "type": "string",
            "example": "#FFFFFF"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSlideRequest": {
        "type": "object",
        "required": ["slide_number", "content"],
        "properties": {
          "slide_number": {
            "type": "integer",
            "example": 1
          },
          "content": {
            "type": "string",
            "example": "# Welcome to SlideBanai"
          },
          "background_color": {
            "type": "string",
            "example": "#FFFFFF"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error", "status"],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/presentations": {
      "get": {
        "tags": ["Presentations"],
        "summary": "List all presentations",
        "operationId": "listPresentations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Results per page",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["draft", "in_progress", "published"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of presentations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Presentation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Presentations"],
        "summary": "Create a new presentation",
        "operationId": "createPresentation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresentationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Presentation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presentation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/presentations/{presentationId}/slides": {
      "get": {
        "tags": ["Slides"],
        "summary": "Get all slides for a presentation",
        "operationId": "listSlides",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of slides",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Slide"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Slides"],
        "summary": "Create a new slide",
        "operationId": "createSlide",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlideRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Slide created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            }
          }
        }
      }
    },
...
  }
}